# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Документация API к приложению достопримечательностей",
        "description": "Авторизация происходит через OAuth2, описание и ссылки на методы надоятся в Authorize",
        "contact": {
            "email": "sergei.klimovich.com"
        },
        "version": "0.1"
    },
    "servers": [
        {
            "url": "https://travel.madskill.ru",
            "description": "API сервер"
        }
    ],
    "paths": {
...
        "/api/attraction": {
            "get": {
                "tags": [
                    "Достопримечательности"
                ],
                "summary": "Одна Достопримечательность",
                "description": "Одна Достопримечательность",
                "operationId": "attraction",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SingleAttraction"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Текст ошибки"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/attractions": {
            "get": {
                "tags": [
                    "Достопримечательности"
                ],
                "summary": "Достопримечательности",
                "description": "Достопримечательности",
                "operationId": "attractions",
                "parameters": [
                    {
                        "name": "latitude",
                        "in": "query",
                        "description": "Широта",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "description": "Долгота",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Attraction"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Текст ошибки"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Пользователь"
                ],
                "summary": "Регистрация",
                "description": "Регистрация",
                "operationId": "registeruser",
                "parameters": [
                    {
                        "name": "name",
                        "in": "header",
                        "description": "Ник пользователя",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "header",
                        "description": "Email пользователя",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "email",
                            "example": "ivanov@ivan.ru"
                        }
                    },
                    {
                        "name": "password",
                        "in": "header",
                        "description": "Пароль пользователя",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное выполнение",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Успешно зарегистрированы"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "string",
                                            "example": "Текст ошибки"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "delete": {
                "tags": [
                    "Пользователь"
                ],
                "summary": "Выход",
                "description": "Выход",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Успешно вышли"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ошибка при выходе"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Attraction": {
                "title": "Attraction",
                "description": "Модель достопримечательности",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "Название",
                        "type": "string",
                        "format": "string",
                        "example": "пример"
                    },
                    "description": {
                        "title": "description",
                        "description": "Описание",
                        "type": "string",
                        "format": "string",
                        "example": "пример"
                    },
                    "image": {
                        "title": "image",
                        "description": "Ссылка на фото",
                        "type": "string",
                        "format": "string",
                        "example": "пример"
                    },
                    "latitude": {
                        "title": "latitude",
                        "description": "Широта",
                        "type": "string",
                        "format": "string",
                        "example": "55.234"
                    },
                    "longitude": {
                        "title": "longitude",
                        "description": "Долгота",
                        "type": "string",
                        "format": "string",
                        "example": "33.234"
                    }
                },
                "type": "object"
            },
            "Photo": {
                "title": "Photo",
                "description": "Модель фотографий",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "attraction_id": {
                        "title": "attraction_id",
                        "description": "attraction_id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "image": {
                        "title": "image",
                        "description": "Название",
                        "type": "string",
                        "format": "string",
                        "example": "пример"
                    }
                },
                "type": "object"
            },
...
            "SingleAttraction": {
                "title": "SingleAttraction",
                "description": "Модель одной достопримечательности",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "Название",
                        "type": "string",
                        "format": "string",
                        "example": "пример"
                    },
                    "description": {
                        "title": "description",
                        "description": "Описание",
                        "type": "string",
                        "format": "string",
                        "example": "пример"
                    },
                    "city": {
                        "title": "city",
                        "description": "Город",
                        "type": "string",
                        "format": "string",
                        "example": "пример"
                    },
                    "photo": {
                        "title": "photo",
                        "description": "photo",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Photo"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Внимание OAuth2 scopes none, grant type code",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "authorizationCode": {
                        "authorizationUrl": "https://travel.madskill.ru/oauth/authorize",
                        "tokenUrl": "https://travel.madskill.ru/oauth/token",
                        "refreshUrl": "https://travel.madskill.ru/token/refresh",
                        "redirectUrl": "https://travel.madskill.ru/auth/callback",
                        "scopes": []
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}